Move pull request update status to buildout.coredev
If a pull request was successful everything was working as expected,
but if there was any test error the update status script did not ran,
as a previous script already failed (bin/test).

Moving the script to buildout.coredev and running it within the same
jenkins build script makes sure that pull requests do get notified.