Removed Python 3.7 from Plone 6.0 jobs.
I already removed them manually in the Jenkins UI, but we don't want them coming back.

I also split the 'Test add-ons on python 3' project in one project for 5.2 and one for 6.0. Use the correct Python versions there.
These add-on jobs are hardly used though, so we could also remove them.
Or advertise them more, I think @gforcada added a comment in an add-on recently.