feat: schedule some core jobs
Rather than running all these core jobs, run them only on a weekly
basis.

The difference between python versions is not that big, and if a job
fails is most of the times due to Jenkins rather than the tests.

This lowers the amount of jobs we have to run for every commit, while
still running it on all version, in case there is indeed a regression on
a certain python version.